scripts/requirements_web.txt
pymongo==4.6.0
python-dotenv==1.0.0
selenium==4.15.2
